requirements.txt
Cython
Pillow
spacy==3.6.1
opencv-python==4.8.0.76
annoy==1.17.3
gensim==4.3.1
